feat: refactor project into modular files
Split logic into separate files:
- driver.py handles main execution flow
- gov_crawler.py contains crawling logic
- utils.py provides helper functions
- config.py stores configuration constants

Improves structure, readability, and maintainability of the project.
