+++ gov_crawler.py
@@ -1,133 +1,39 @@
-#!/usr/bin/env python3
-"""
-Cleaned crawler: yohananof (login) + king-like providers (kingstore/maayan).
-Downloads latest 2 price + latest 2 promo .gz files per provider.
-"""
-
-import os
+# gov_crawler.py
 import re
 import time
-import logging
-import certifi
-import requests
-from datetime import datetime, timedelta
+from datetime import datetime
 from urllib.parse import urljoin, urlparse, unquote
-
-from selenium import webdriver
-from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-from selenium.common.exceptions import TimeoutException, WebDriverException
+from selenium.common.exceptions import TimeoutException
 
-import boto3  # ADDED for AWS S3 upload
+from config import (
+    BASE_LISTING_URL,
+    PAGE_WAIT,
+    TARGET_STORES,
+    YOHANANOF_USERNAME,
+    YOHANANOF_PASSWORD,
+    VERIFY_SSL,
+    logger,
+    GZ_ONCLICK_RE,
+    S3_CLIENT,
+    BUCKET_NAME,
+)
+from utils import (
+    ensure_dir,
+    parse_absolute_he_date,
+    parse_relative_he,
+    session_from_driver,
+    download_stream,
+    filename_from_url,
+)
 
-# ---------- CONFIG ----------
-BASE_LISTING_URL = "https://www.gov.il/he/pages/cpfta_prices_regulations"
-DOWNLOAD_DIR = os.path.join(os.getcwd(), "providers")
-HEADLESS = True
-VERIFY_SSL = True
-
-TARGET_STORES = [
-    ("מ. יוחננוף", "yohananof"),
-    ("אלמשהדאוי קינג סטור", "kingstore"),
-    ("ג.מ מעיין אלפיים", "maayan"),
-]
-
-YOHANANOF_USERNAME = "yohananof"
-YOHANANOF_PASSWORD = ""
-
-PAGE_WAIT = 0.6
-
-# AWS S3 CONFIG 
-BUCKET_NAME = "govil-price-lists" 
-
-# Create the S3 client 
-s3_client = boto3.client("s3")
-
-logging.basicConfig(level=logging.INFO, format="[%(asctime)s] %(levelname)s: %(message)s")
-logger = logging.getLogger("crawler")
-
-# regexes
-ABS_DATE_RE = re.compile(r"(\d{1,2}):(\d{2})\s+(\d{1,2})/(\d{1,2})/(\d{4})")
-REL_HE_RE = re.compile(r"לפני\s*(\d+)?\s*(שנייה|שניות|דקה|דקות|שעה|שעות|יום|ימים)", re.I)
-GZ_ONCLICK_RE = re.compile(r"Download\(['\"]([^'\"]+\.gz)['\"]\)", re.I)
-
-# ---------- helpers ----------
-def ensure_dir(p):
-    os.makedirs(p, exist_ok=True)
-
-def parse_absolute_he_date(text):
-    if not text:
-        return None
-    m = ABS_DATE_RE.search(text)
-    if not m:
-        return None
-    hh, mm, dd, mo, yy = m.groups()
-    try:
-        return datetime(int(yy), int(mo), int(dd), int(hh), int(mm))
-    except Exception:
-        return None
-
-def parse_relative_he(text):
-    if not text:
-        return None
-    m = REL_HE_RE.search(text)
-    if not m:
-        return None
-    num = m.group(1)
-    unit = m.group(2)
-    n = int(num) if num and num.isdigit() else 1
-    now = datetime.now()
-    unit = unit.strip()
-    if unit.startswith("שנייה"): return now - timedelta(seconds=n)
-    if unit.startswith("דקה"): return now - timedelta(minutes=n)
-    if unit.startswith("שעה"):  return now - timedelta(hours=n)
-    if unit.startswith("יום"):   return now - timedelta(days=n)
-    return now
-
-def session_from_driver(driver):
-    s = requests.Session()
-    ua = driver.execute_script("return navigator.userAgent;")
-    s.headers.update({"User-Agent": ua})
-    for c in driver.get_cookies():
-        s.cookies.set(c['name'], c['value'], domain=c.get('domain', None), path=c.get('path', '/'))
-    return s
-
-def download_stream(session, url, dest, verify=True):
-    try:
-        verify_target = certifi.where() if verify else False
-        with session.get(url, stream=True, timeout=60, verify=verify_target) as r:
-            r.raise_for_status()
-            with open(dest, "wb") as fh:
-                for chunk in r.iter_content(8192):
-                    if chunk:
-                        fh.write(chunk)
-        return True
-    except Exception as e:
-        logger.error(f"Failed to download {url}: {e}")
-        return False
-
-# ---------- Selenium ----------
-def make_driver(headless=HEADLESS):
-    opts = Options()
-    if headless:
-        opts.add_argument("--headless=new")
-    opts.add_argument("--window-size=1400,900")
-    opts.add_argument("--disable-gpu")
-    opts.add_argument("--no-sandbox")
-    opts.add_argument("--disable-dev-shm-usage")
-    opts.add_experimental_option("excludeSwitches", ["enable-automation"])
-    opts.add_experimental_option("useAutomationExtension", False)
-    driver = webdriver.Chrome(options=opts)
-    driver.set_page_load_timeout(45)
-    return driver
-
-# ---------- Crawler ----------
 class GovCrawler:
     def __init__(self, driver):
         self.driver = driver
 
+    # listing page
     def load_listing_rows(self):
         logger.info("Loading gov.il listing page")
         self.driver.get(BASE_LISTING_URL)
@@ -175,14 +81,19 @@ class GovCrawler:
     def process_store(self, fragment, folder):
         logger.info(f"Processing: {fragment} -> {folder}")
         name, anchor, row = self.find_store_anchor(fragment)
-        if not name or not anchor:
-            logger.warning(f"Store link not found for fragment '{fragment}'")
+        if not name:
+            logger.warning(f"Store fragment {fragment} not found")
+            return
+        if not anchor:
+            logger.warning(f"No listing link for {name}")
             return
         if not self.open_provider(anchor):
             logger.warning("Could not open provider page")
             return
 
-        ensure_dir(os.path.join(DOWNLOAD_DIR, folder))
+        ensure_dir(f"{BUCKET_NAME}")  # no-op if already exists locally; kept for parity
+        ensure_dir(__import__("os").path.join(__import__("os").getcwd(), "providers", folder))
+
         try:
             if "יוחננוף" in name:
                 self.handle_yohananof(folder)
@@ -191,6 +102,7 @@ class GovCrawler:
         except Exception as e:
             logger.exception(f"Error handling provider {folder}: {e}")
         finally:
+            # close provider tab and return
             try:
                 if len(self.driver.window_handles) > 1:
                     self.driver.close()
@@ -198,7 +110,7 @@ class GovCrawler:
             except Exception:
                 pass
 
-    # ----- Yohananof -----
+    # Yohananof flow
     def handle_yohananof(self, folder):
         logger.info("Yohananof: attempting login")
         try:
@@ -213,12 +125,13 @@ class GovCrawler:
         except TimeoutException:
             logger.debug("Login form not present (maybe already logged in)")
 
+        # click a post-login "Price" link if present
         clicked = False
         possible_texts = ["Price", "Prices", "מחירים", "לצפייה במחירים", "PriceFull"]
-        for text in possible_texts:
+        for t in possible_texts:
             try:
                 a = WebDriverWait(self.driver, 4).until(
-                    EC.element_to_be_clickable((By.XPATH, f"//a[contains(normalize-space(.),'{text}')]"))
+                    EC.element_to_be_clickable((By.XPATH, f"//a[contains(normalize-space(.),'{t}')]"))
                 )
                 try:
                     a.click()
@@ -252,7 +165,7 @@ class GovCrawler:
             return
 
         sess = session_from_driver(self.driver)
-        # optional: CSRF from meta
+        # optional CSRF token
         try:
             meta = self.driver.find_element(By.XPATH, "//meta[@name='csrftoken']")
             token = meta.get_attribute("content")
@@ -263,9 +176,9 @@ class GovCrawler:
 
         self.select_and_download_from_rows(rows, folder, sess)
 
-    # ----- King-like providers -----
+    # Generic king-like providers
     def handle_kinglike(self, folder):
-        logger.info("Generic king-like provider: trying to trigger BuildHtml / Download buttons")
+        logger.info("Generic king-like provider: attempt to trigger BuildHtml / Download buttons")
         try:
             buttons = self.driver.find_elements(By.XPATH, "//button[contains(normalize-space(.),'הורדה') or contains(normalize-space(.),'להורדה')]")
             for b in buttons[:6]:
@@ -286,13 +199,13 @@ class GovCrawler:
             rows = self.extract_gz_from_html(self.driver.page_source or "", base=self.driver.current_url)
 
         if not rows:
-            logger.warning("Generic provider: no .gz files found")
+            logger.warning("Generic: no files found on page")
             return
 
         sess = session_from_driver(self.driver)
         self.select_and_download_from_rows(rows, folder, sess)
 
-    # ----- extraction helpers -----
+    # extraction helpers
     def extract_table_rows_with_gz(self, relative_time=False):
         results = []
         try:
@@ -309,7 +222,7 @@ class GovCrawler:
                     try:
                         a = r.find_element(By.XPATH, ".//a[contains(@href,'.gz')]")
                         href = a.get_attribute("href")
-                        fname = unquote(os.path.basename(urlparse(href).path))
+                        fname = unquote(urlparse(href).path.split("/")[-1])
                     except Exception:
                         continue
                 type_text = tds[2].text.strip() if len(tds) >= 3 else ""
@@ -320,6 +233,7 @@ class GovCrawler:
                 else:
                     dt = parse_absolute_he_date(date_text)
                 if dt is None:
+                    # fallback parse in filename
                     m = re.search(r"(20\d{2})(\d{2})(\d{2})(\d{2})?(\d{2})?", fname)
                     if m:
                         try:
@@ -330,6 +244,7 @@ class GovCrawler:
                         except Exception:
                             dt = datetime.now()
                 dt = dt or datetime.now()
+                # download hint (onclick or anchor)
                 download_hint = None
                 try:
                     btn = r.find_element(By.XPATH, ".//button[@onclick]")
@@ -346,6 +261,7 @@ class GovCrawler:
                         download_hint = href
                 except Exception:
                     pass
+
                 results.append({"filename": fname, "type": type_text, "date": dt, "download_hint": download_hint})
             except Exception:
                 continue
@@ -355,7 +271,7 @@ class GovCrawler:
         found = []
         for m in re.finditer(r'href=[\'"]([^\'"]+?\.gz)[\'"]', html, re.I):
             url = urljoin(base, m.group(1))
-            fn = unquote(os.path.basename(urlparse(url).path))
+            fn = unquote(urlparse(url).path.split("/")[-1])
             found.append({"filename": fn, "type": "", "date": datetime.now(), "download_hint": url})
         for m in GZ_ONCLICK_RE.finditer(html):
             fn = m.group(1)
@@ -363,10 +279,11 @@ class GovCrawler:
             found.append({"filename": fn, "type": "", "date": datetime.now(), "download_hint": fn})
         return found
 
-    # ----- select & download -----
+    # select & download
     def select_and_download_from_rows(self, rows, folder, session):
-        logger.info(f"Found {len(rows)} candidates for {folder}")
-        price, promo = [], []
+        logger.info(f"Found {len(rows)} candidates")
+        price = []
+        promo = []
         for r in rows:
             key = (r.get("filename", "") + " " + r.get("type", "")).lower()
             if "promo" in key or "מבצע" in key or "מבצעים" in key or "promo" in r.get("filename", "").lower():
@@ -375,13 +292,15 @@ class GovCrawler:
                 price.append(r)
         price.sort(key=lambda x: x.get("date") or datetime.min, reverse=True)
         promo.sort(key=lambda x: x.get("date") or datetime.min, reverse=True)
-        price = price[:2]; promo = promo[:2]
-        logger.info(f"Will download {len(price)} price + {len(promo)} promo for {folder}")
+        price = price[:2]
+        promo = promo[:2]
+        logger.info(f"Will download {len(price)} price + {len(promo)} promo files for {folder}")
 
         for item in price + promo:
             final_url = None
             hint = item.get("download_hint")
             fn = item.get("filename")
+            # if hint is full url
             if hint and isinstance(hint, str) and hint.lower().startswith("http"):
                 final_url = hint
             elif hint and isinstance(hint, str) and hint.lower().endswith(".gz"):
@@ -389,9 +308,9 @@ class GovCrawler:
                 parsed = urlparse(base)
                 base_root = f"{parsed.scheme}://{parsed.netloc}"
                 download_ajax = urljoin(base_root, "Download.aspx?FileNm=" + hint)
-                logger.info(f"Requesting Download.aspx for {fn}: {download_ajax}")
+                logger.info(f"Requesting provider Download.aspx for {fn}: {download_ajax}")
                 try:
-                    r = session.post(download_ajax, timeout=20, verify=certifi.where() if VERIFY_SSL else False)
+                    r = session.post(download_ajax, timeout=20, verify= (True if VERIFY_SSL else False))
                     if r.status_code == 200:
                         try:
                             data = r.json()
@@ -411,6 +330,7 @@ class GovCrawler:
                 except Exception as e:
                     logger.warning(f"Download.aspx call failed for {hint}: {e}")
             else:
+                # fallback try to find anchor in DOM
                 try:
                     a = self.driver.find_element(By.XPATH, f"//a[contains(@href, '{fn}')]")
                     href = a.get_attribute("href")
@@ -423,44 +343,21 @@ class GovCrawler:
                 logger.warning(f"Could not resolve final URL for {fn}; skipping")
                 continue
 
-            ensure_dir(os.path.join(DOWNLOAD_DIR, folder))
-            dest = os.path.join(DOWNLOAD_DIR, folder, fn)
+            ensure_dir(__import__("os").path.join(__import__("os").getcwd(), "providers", folder))
+            dest = __import__("os").path.join(__import__("os").getcwd(), "providers", folder, fn)
             logger.info(f"Downloading {final_url} -> {dest} (date={item.get('date')})")
             ok = download_stream(session, final_url, dest, verify=VERIFY_SSL)
             if not ok and VERIFY_SSL:
-                logger.warning("Retrying download with verify=False")
+                logger.warning("Retrying with verify=False")
                 ok = download_stream(session, final_url, dest, verify=False)
 
-            # === Upload to S3 ===
+            # upload to S3 if successful
             if ok:
                 s3_key = f"{folder}/{fn}"
                 try:
-                    s3_client.upload_file(dest, BUCKET_NAME, s3_key)
+                    S3_CLIENT.upload_file(dest, BUCKET_NAME, s3_key)
                     logger.info(f"Uploaded to s3://{BUCKET_NAME}/{s3_key}")
                 except Exception as e:
                     logger.error(f"Failed to upload {fn} to S3: {e}")
-            # ====================
 
-            time.sleep(0.25)
-
-# ---------- main ----------
-def main():
-    ensure_dir(DOWNLOAD_DIR)
-    driver = make_driver()
-    crawler = GovCrawler(driver)
-    try:
-        for frag, folder in TARGET_STORES:
-            try:
-                crawler.process_store(frag, folder)
-            except Exception as e:
-                logger.exception(f"Store {frag} failed: {e}")
-            time.sleep(0.9)
-    finally:
-        try:
-            driver.quit()
-        except Exception:
-            pass
-    logger.info("All done.")
-
-if __name__ == "__main__":
-    main()
+            time.sleep(0.35)
